requirements.txt
# Municipality RAG - Python Dependencies

# Core LLM and RAG libraries
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.6

# Vector database
chromadb==0.4.22

# Embeddings and LLM APIs
openai==1.12.0              # OpenAI API (optional)
anthropic==0.18.1           # Anthropic Claude API (optional)
ollama==0.1.7               # Ollama Python client (recommended for free local LLMs)

# Data processing
pyyaml==6.0.1               # YAML parsing
python-dotenv==1.0.1        # Environment variable management
pydantic==2.6.1             # Data validation

# Web UI (optional)
streamlit==1.31.0           # Web interface for RAG system

# Utilities
requests==2.31.0            # HTTP requests
tqdm==4.66.1                # Progress bars

# Testing
pytest==8.0.0               # Testing framework
pytest-cov==4.1.0           # Coverage reports

# Jupyter notebooks (optional)
jupyter==1.0.0
ipykernel==6.29.0

# Document processing (if handling PDFs/Word docs)
pypdf==4.0.1                # PDF reading (optional)
python-docx==1.1.0          # Word doc reading (optional)

# Logging and monitoring
loguru==0.7.2               # Better logging

# Development tools
black==24.1.1               # Code formatter
flake8==7.0.0               # Linter
mypy==1.8.0                 # Type checker
